More of scan code added
The commit adds more code related to scanning and parsing Lua data. It includes changes to the `add_SubNodeToSubnode` method in the `LuaNode` class, as well as modifications to the `generateNodeTree` method in the `TreeStorage` class. The changes involve handling different levels of indentation, extracting node names and content, and adding nodes to the tree structure.