Now uses file data to create Node Tree storage
Now uses file data to create Node Tree storage

The commit updates the `TreeStorage` class in `krangler.py` to now accept `fileData` as a parameter in its constructor. If `fileData` is not empty, it calls the `generateNodeTree()` method to generate the node tree based on the provided data. This change allows for more flexibility and customization when initializing the `TreeStorage` object.