adding list info scanning
This commit adds functionality to scan and process list information in the code. It introduces a new scanning level called "listInfo" which allows for the detection of elements inside list variables. When a comma or newline is encountered, it adds a list node to the appropriate subnode or top-level key. This change enhances the overall parsing capabilities of the code.