Corrected recursive function used by functions
The commit corrects the recursive function used by functions in the code. The use of `printDebugInfo` and `printNodeContent` functions have been replaced with `printNodeContentFromSubnode`. This change ensures that the correct subnode is used for recussive functions (subnodes only used for 1st level subnodes and not the children of 1st level subnodes)